+++ kite/src/main/kotlin/com/tang/kite/wrapper/delete/DeleteWrapper.kt
@@ -2,8 +2,10 @@ package com.tang.kite.wrapper.delete
 
 import com.tang.kite.config.SqlConfig
 import com.tang.kite.constants.SqlString.DELETE_FROM
+import com.tang.kite.enumeration.SqlType
 import com.tang.kite.mapper.BaseMapper
 import com.tang.kite.sql.SqlStatement
+import com.tang.kite.utils.Fields
 import com.tang.kite.utils.Reflects
 import com.tang.kite.wrapper.Wrapper
 import com.tang.kite.wrapper.statement.LogicalStatement
@@ -15,6 +17,8 @@ import com.tang.kite.wrapper.where.AbstractWhereWrapper
  * @author Tang
  */
 class DeleteWrapper<T> : AbstractWhereWrapper<DeleteWhereWrapper<T>, T>, Wrapper<T> {
+
+    private lateinit var tableClass: Class<T>
 
     private lateinit var table: String
 
@@ -67,10 +71,17 @@ class DeleteWrapper<T> : AbstractWhereWrapper<DeleteWhereWrapper<T>, T>, Wrapper
     }
 
     fun setTableClassIfNotSet(clazz: Class<T>) {
-        if (::table.isInitialized) {
+        if (::tableClass.isInitialized) {
             return
         }
+        this.tableClass = clazz
         this.table = Reflects.getTableName(clazz)
+    }
+
+    fun setTableFillFields() {
+        Fields.setTableFillFields(tableClass, SqlType.DELETE) { column, value ->
+            deleteWhereWrapper.eq(column, value)
+        }
     }
 
     override fun getSqlStatement(): SqlStatement {