+++ kite/src/main/kotlin/com/tang/kite/wrapper/query/QuerySelectWrapper.kt
@@ -2,7 +2,9 @@ package com.tang.kite.wrapper.query
 
 import com.tang.kite.constants.SqlString.COMMA_SPACE
 import com.tang.kite.constants.SqlString.FROM
+import com.tang.kite.enumeration.SqlType
 import com.tang.kite.function.SFunction
+import com.tang.kite.utils.Fields
 import com.tang.kite.utils.Reflects
 import com.tang.kite.wrapper.Column
 import kotlin.reflect.KMutableProperty1
@@ -135,6 +137,12 @@ class QuerySelectWrapper<T>(
         }
         this.tableClass = clazz
         this.table = Reflects.getTableName(clazz)
+    }
+
+    fun setTableFillFields() {
+        Fields.setTableFillFields(tableClass, SqlType.SELECT) { column, value ->
+            queryWrapper.queryWhereWrapper.eq(column, value)
+        }
     }
 
     /**